+++ .github/workflows/rule-validate.yml
@@ -114,32 +114,22 @@ jobs:
         with:
           python-version: '3.10'
 
-      - name: Checkout scripts from Sublime fork main
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
-        with:
-          repository: sublime-security/sublime-rules
-          ref: main
-          path: sublime-rules-main
-
       - name: Check for invisible characters
         if: github.event_name != 'issue_comment'
-        run: python sublime-rules-main/scripts/check_invisible_chars.py
+        run: python .github/scripts/check_invisible_chars.py
 
       - name: Add Rule IDs as Needed & Check for Duplicates
         if: github.event_name != 'issue_comment'
         # Run before testing, just in case this could invalidate the rule itself
         run: |
-          pip install -r sublime-rules-main/scripts/generate-rule-ids/requirements.txt
-          python sublime-rules-main/scripts/generate-rule-ids/main.py
+          pip install -r .github/scripts/generate_rule_ids/requirements.txt
+          python .github/scripts/generate_rule_ids/main.py
 
       - name: Auto-format MQL
         if: github.event_name != 'issue_comment'
         run: |
           pip install -q requests pyyaml
-          python sublime-rules-main/scripts/mql_format.py detection-rules/*.yml insights/**/*.yml
-
-          # Delete path to prevent interference with later steps (such as git add and commit)
-          rm -r sublime-rules-main
+          python .github/scripts/mql_format.py detection-rules/*.yml insights/**/*.yml
 
       - name: Validate Rules
         if: github.event_name != 'issue_comment'
